Proposed Fix for Skill Name Mapping Issue
Skill names are failing to populate, due to a mismatch in the named
property "Name" in the Skill class.  The REST response from LinkedIn
returns a nested Skill node, which in turn, contains the name.  Proposed
modifications are to rename the Skill class as SkillItem so that the
"Name" property in that class can be renamed to "Skill", in order for it
to populate.